requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
Django==5.0.7
djangorestframework==3.15.2
minio==7.2.7
pycparser==2.22
pycryptodome==3.20.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
